Shrink DbgVerbose to int8_t
Valid values are in [-3, 2].

+++ common/include/args.h
@@ -92,7 +92,7 @@ struct CArg : prohibit_void_ptr<CArg>
 	bool DbgGlRGBA2Ok;
 #endif
 	uint8_t DbgBpp;
-	int DbgVerbose;
+	int8_t DbgVerbose;
 	int SysMaxFPS;
 	std::string SysMissionDir;
 };
